refactor(evaluate/utils): move metric_list_to_json into scoring.py, delete io_utils.py
